+++ lib/rubocop/cop/style/tab.rb
@@ -1,5 +1,7 @@
 # encoding: utf-8
 
+require 'set'
+
 module RuboCop
   module Cop
     module Style
@@ -8,12 +10,14 @@ module RuboCop
         MSG = 'Tab detected.'
 
         def investigate(processed_source)
+          str_lines = string_literal_lines(processed_source.ast)
+
           processed_source.lines.each_with_index do |line, index|
             match = line.match(/^( *)[\t ]*\t/)
             next unless match
+            next if str_lines.include?(index + 1)
 
             spaces = match.captures[0]
-
             range = source_range(processed_source.buffer,
                                  index + 1,
                                  (spaces.length)...(match.end(0)))
@@ -27,6 +31,24 @@ module RuboCop
         def autocorrect(range)
           lambda do |corrector|
             corrector.replace(range, range.source.gsub(/\t/, '  '))
+          end
+        end
+
+        def string_literal_lines(ast)
+          # which lines start inside a string literal?
+          return [] if ast.nil?
+
+          ast.each_node(:str, :dstr).each_with_object(Set.new) do |str, lines|
+            loc = str.location
+
+            str_lines = if loc.is_a?(Parser::Source::Map::Heredoc)
+                          body = loc.heredoc_body
+                          (body.first_line)..(body.last_line)
+                        else
+                          (loc.first_line + 1)..(loc.last_line)
+                        end
+
+            lines.merge(str_lines.to_a)
           end
         end
       end